requirements.txt
anthropic
openai
requests
beautifulsoup4
moviepy
Pillow
selenium
webdriver-manager
peewee
dotenv